Fix usage of Qt inside the `electroncash` module
Qt usage is not allowed in the `electroncash/` module. We removed all
assumptions that the app is running under Qt for the lns.py file, as
well as contacts.py, both of which life in `electroncash/`.
